feat(cli): gazetteer build admin|candidate + turnkey build + structural verify; wof namespace retired
One artifact→verb namespace: gazetteer build (full chain: verified sealed admin → candidate), build admin, build candidate (the old bare build), verify (the standalone #1026/#1015 gate — exits non-zero; live-DB smoke correctly FAILS node-census on the flattened countries with 15/15 reverse), inspect tree|graph|mermaid|sync (ex-wof). wof/* are one-minor-version deprecation shims; the stale partial wof prepare is deleted (superseded by build admin).
