+++ mailwoman/commands/gazetteer/publish.tsx
@@ -21,7 +21,12 @@ import { Box, Text } from "ink"
 import { useEffect, useState } from "react"
 import zod from "zod"
 
-import { DEFAULT_CANDIDATE_OUT, defaultGazetteerVersion, publishGazetteer, wofDir } from "../../gazetteer-pipeline/index.js"
+import {
+	DEFAULT_CANDIDATE_OUT,
+	defaultGazetteerVersion,
+	publishGazetteer,
+	wofDir,
+} from "../../gazetteer-pipeline/index.js"
 import type { CommandComponent } from "../../sdk/cli.js"
 
 const ArgumentsSchema = zod.array(
